requirements.txt
flask_cors==5.0.1
Flask_Minify==0.50
itsdangerous==2.2.0
python-dotenv==1.1.0
SQLAlchemy==2.0.40
